+++ Zbang.Cloudents.Mvc4WebRole/Js/components/flashcard/flashcardCreate.controller.ts
@@ -11,6 +11,7 @@ module app {
             this.cards = [];
             this.id = input.id;
             this.name = input.name;
+            input.cards = input.cards || [];
             for (let i = 0; i < input.cards.length; i++) {
                 this.cards.push(new Card().deserialize(input.cards[i]));
             }
@@ -45,7 +46,7 @@ module app {
         image: string;
         deserialize(input: CardSlide) {
             this.text = input.text;
-            this.image = input.image;// || "http://lorempixel.com/400/200/";
+            this.image = input.image;
             return this;
         }
         uploadProgress: number;
@@ -102,6 +103,9 @@ module app {
         publish() {
             var self = this;
             if (this.form.$invalid) {
+                if ((this.form["name"] as angular.INgModelController).$invalid) {
+                    (this.form["name"] as angular.INgModelController).$setTouched();
+                }
                 return;
             } else {
                 if (!this.data.id) {
@@ -127,7 +131,7 @@ module app {
         }
         create() {
             const self = this;
-            
+
             function afterCall() {
                 self.serviceCalled = false;
                 self.disabled = false;
@@ -172,7 +176,7 @@ module app {
             this.data.cards.push(new Card());
         }
         close(ev) {
-            if (!this.data.id) {
+            if (!this.data.id && !this.form.$dirty) {
                 this.navigateBackToBox();
                 return;
             }
@@ -185,7 +189,11 @@ module app {
                 .cancel(this.resManager.get('quizSaveAsDraft'));
 
             this.$mdDialog.show(confirm).then(() => {
-                this.flashcardService.delete(this.data.id).then(this.navigateBackToBox);
+                if (this.data.id) {
+                    this.flashcardService.delete(this.data.id).then(this.navigateBackToBox);
+                } else {
+                    this.navigateBackToBox();
+                }
             }, () => {
                 this.create().then(this.navigateBackToBox);
             });
@@ -213,7 +221,9 @@ module app {
                         ]
                     },
                     resize: {
-                        preserve_headers: false
+                        preserve_headers: false,
+                        width: 350,
+                        height: 350
                     }
                 }
             },
